Simplistic focusable mechanism now in place.
Events bubble down and if the click down position is in the bounds of the view
the view is focused and then unfocused on the click up. This is to be changed
to the Android model of different modes such as focus mode and touch mode. The
two modes will be keyboard mode and touch (click) mode.