Advances on focusable mechanism, to include different modes (TOUCH_MODE, KEY_MODE).
However, the differentiation between the modes and the dispatching events to
the children view is not complete. Currently, when the main menu view is clicked
it is able to accept key input. The other menus do not, and the game is able to
be played properly once the option is selected.